edge-llm-eval: real capability matrix on Qwen2.5-0.5B
First working capability matrix across FP16 / INT8 / INT4 (bnb NF4):
- FP16 ppl 31.770
- INT8 ppl 31.790  (+0.06% — essentially lossless)
- INT4 ppl 40.332  (+26.9% — meaningful quality drop)

Two bugs found and fixed along the way, both documented in the report:

1. Generative regex scorer gave a nonsense inversion (FP16=0, INT4=9/20)
   because a 0.5B base model can't do (a+b)*c without CoT and the regex
   was picking up intermediate numbers. Switched to logprob-scored MCQ.

2. Even after (1), FP16 and INT8 reported *identical* perplexity because
   log-probs were being summed in fp16, snapping to fp16's 0.25-wide
   precision grid at magnitude ~300. Upcast logits to fp32 before
   log_softmax and accumulate in fp64 — now the 0.06% gap is visible.

Also:
- arith_mcq mean_margin tracks quant quality cleanly even when accuracy
  is too noisy at n=30 (0.85 -> 0.94 -> 0.68)
- toy_mcq margin *increases* with aggressive quant (3.73 -> 3.81 -> 4.01),
  flagged as surprising-and-noted-but-not-explained

Adds: plot.py, docs/report.md, results/capability_matrix.{json,png}

+++ edge-llm-eval-harness/src/harness.py
@@ -68,10 +68,16 @@ class HFBackend:
         prompt_len = self._tok(prompt, return_tensors="pt").input_ids.shape[1]
         with torch.no_grad():
             logits = self._model(**ids).logits[0]
+        # Upcast to fp32 before log_softmax so the accumulated sum of
+        # token log-probs doesn't get snapped to fp16's coarse grid at
+        # large magnitudes (where fp16 precision is ~0.25). Without this,
+        # fp16 and int8 backends report *identical* logprobs even though
+        # their logits differ, which makes the whole eval look insensitive.
+        logits = logits.float()
         log_probs = torch.log_softmax(logits, dim=-1)
         cont_ids = ids.input_ids[0, prompt_len:]
         token_lps = log_probs[prompt_len - 1 : -1].gather(-1, cont_ids[:, None]).squeeze(-1)
-        return float(token_lps.sum().item())
+        return float(token_lps.double().sum().item())
 
 
 def time_generate(backend: Backend, prompt: str, max_new_tokens: int = 64):
